Function ic_state_manager::checkpoint::make_checkpoint

pub fn make_checkpoint(
    state: &ReplicatedState, 
    height: Height, 
    layout: &StateLayout
) -> Result<ReplicatedState, CheckpointError>

Creates a checkpoint of the node state using specified directory layout. Returns a new state that is equivalent the to given one and a result of the operation.

If the result is Ok, the returned state is "rebased" to use files from the newly created checkpoint. If the result is Err, the returned state is exactly the one that was passed as argument.