Module ic_state_manager::manifest

Modules

hash

Structs

DiffScript
A script describing how to turn an old state into a new state.

Enums

ChunkValidationError

ManifestValidationError

Constants

CURRENT_STATE_SYNC_VERSION
The version of StateSync protocol that should be used for all newly produced states.

DEFAULT_CHUNK_SIZE

STATE_SYNC_V1

Functions

compute_manifest
Computes manifest for the checkpoint located at checkpoint_root_path.

diff_manifest
Computes diff between two manifests and get DiffScript.

filter_out_zero_chunks
Filters out all-zero chunks in the manifest chunk table and returns the set of remaining chunks indices.

manifest_hash
Computes root hash of the manifest. See note [Manifest Hash].

validate_chunk
Checks that the size and hash of the received chunk match the chunk table of the manifest.

validate_manifest
Validates manifest contents and checks that the hash of the manifest matches the expected root hash.

Type Definitions

NewIndex

OldIndex