Module ic_state_manager::stream_encoding

Functions

decode_labeled_tree
Decodes a labeled tree from a byte buffer.

decode_slice_from_tree
Recovers a stream slice from its canonical form.

decode_stream_slice
Decodes a stream slice and the subnet it came from from a serialized canonical tree.

encode_stream_slice
Encodes a stream slice for the specified subnet, consisting of available messages beginning at from and ending before to, of total size at most byte_limit, using canonical tree form.

encode_tree

stream_slice_partial_tree
Creates a partial tree (structure only, empty values) in canonical tree form (see encode_stream_slice) for a stream slice for the specified subnet, with messages beginning at from (inclusive) and ending at to (exclusive), to be used for witness generation.