Function ic_state_manager::stream_encoding::encode_stream_slice

pub fn encode_stream_slice(
    state: &ReplicatedState, 
    subnet: SubnetId, 
    from: StreamIndex, 
    to: StreamIndex, 
    byte_limit: Option<usize>
) -> (LabeledTree<Vec<u8>>, StreamIndex)

Encodes a stream slice for the specified subnet, consisting of available messages beginning at from and ending before to, of total size at most byte_limit, using canonical tree form.

Returns the encoded slice and the actual end index (which may be different from to if byte_limit is reached).

For example, the slice for subnet 25 with messages [5, 10) (assuming the full range of messages in the stream is [3, 20)) will look like this:

*
| streams
`---------*
          | 25
          `----*
               | header
               +-------- { begin: 3, end: 20, signals_end: ... }
               | messages
               `---------*
                         | 5
                         +---- { blob }
                         | 6
                         +---- { blob }
                         |
                         …
                         | 9
                         +---- { blob }

See ic_replicated_state::canonical::traversal for more details on the tree structure.