Struct ic_state_manager::StateManagerImpl

pub struct StateManagerImpl { /* fields omitted */ }

Implementations

impl StateManagerImpl

pub fn new(
verifier: Arc<dyn Verifier>,
own_subnet_id: SubnetId,
own_subnet_type: SubnetType,
log: ReplicaLogger,
metrics_registry: &MetricsRegistry,
config: &Config,
malicious_flags: MaliciousFlags
) -> Self

pub fn state_layout(&self) -> &StateLayout

Returns StateLayout pointing to the directory managed by this StateManager.

pub fn create_chunkable_state(
&self,
id: &StateSyncArtifactId
) -> Box<dyn Chunkable + Send + Sync>

Returns requested state as a Chunkable artifact for StateSync.