Module ic_state_manager::tree_hash

Structs

HashingVisitor
A visitor that constructs a hash tree by traversing a replicated state.

Functions

hash_partial_state
Compute the hash tree corresponding to the partial state.

hash_state
Compute the hash tree corresponding to the full replicated state.